refactor(#1459): relocate cache modules to cache/ subpackage + CacheCoordinator (Phase 5)
Move 5 cache modules into the cache/ subpackage with backward-compat shims:
- rebac_cache.py -> cache/result_cache.py (1,519 LOC, L1 permission cache)
- permission_boundary_cache.py -> cache/boundary.py (451 LOC)
- dir_visibility_cache.py -> cache/visibility.py (396 LOC)
- rebac_iterator_cache.py -> cache/iterator.py (319 LOC)
- leopard.py -> cache/leopard.py (798 LOC, transitive group closure)

Create cache/coordinator.py (382 LOC): unified CacheCoordinator that
orchestrates invalidation across all cache layers (L1, boundary,
visibility, iterator, zone graph). Wired into EnhancedReBACManager
with callback forwarding for boundary and visibility invalidators.

All existing import paths continue to work via backward-compat shims.
